requirements.txt
streamlit
pandas
numpy
PyMuPDF
pdfplumber
python-docx
spacy
nltk
scikit-learn
transformers
torch
